Crate maidsafe_utilities

MaidSafe Utilities

Rust utility functions provided by MaidSafe.

Modules

event_sender
Utilities related to event-subsetting.

log
Allows initialising the env_logger with a standard message format.

serialisation
Functions for serialisation and deserialisation

thread
Utilities related to threading.

Macros

thread!
This macro is intended to be used in all cases where we want to spawn a new thread of execution and if that is not possible then panic out.

unwrap_option!
A replacement for calling unwrap() on an Option.

unwrap_result!
A replacement for calling unwrap() on a Result.