Module maidsafe_utilities::event_sender

Utilities related to event-subsetting.

Structs

EventSender
This structure is coded to achieve event-subsetting. Receivers in Rust are blocking. One cannot listen to multiple receivers at the same time except by using try_recv which again is bad for the same reasons spin-lock based on some sleep is bad (wasting cycles, 50% efficient on an average etc.).

Enums

EventSenderError
Errors that can be returned by EventSender

MaidSafeEventCategory
Category of events meant for a MaidSafe observer listening to both, routing and crust events

Type Definitions

MaidSafeObserver
Observer that Crust (and users of Routing if required) must allow to be registered