Function maidsafe_utilities::log::init

pub fn init(show_thread_name: bool)

This function initialises the env_logger.

An example of a log message is:

W 19:33:49.245434 <main> [example:src/main.rs:50] Warning level message.
^        ^          ^        ^          ^                    ^
|    timestamp      | top-level module  |                 message
|                   |                   |
|              thread name       file and line no.
|
level (E, W, I, D, or T for error, warn, info, debug or trace respectively)

Logging of the thread name is enabled or disabled via the show_thread_name parameter. If enabled, and the thread executing the log statement is unnamed, the thread name is shown as ???.

The function can safely be called multiple times concurrently.

Examples

#[macro_use]
extern crate log;
#[macro_use]
extern crate maidsafe_utilities;

fn main() {
    maidsafe_utilities::log::init(true);

    warn!("A warning");

    let unnamed = ::std::thread::spawn(move || info!("Message in unnamed thread"));
    let _ = unnamed.join();

    let _named = ::maidsafe_utilities::thread::RaiiThreadJoiner::new(thread!("Worker",
                     move || error!("Message in named thread")));

    // W 12:24:07.064746 <main> [example:src/main.rs:9] A warning
    // I 12:24:07.065746 ??? [example:src/main.rs:11] Message in unnamed thread
    // E 12:24:07.065746 Worker [example:src/main.rs:14] Message in named thread
}