Function maidsafe_utilities::log::init_to_file

pub fn init_to_file<P: AsRef<Path>>(show_thread_name: bool, file_path: P) -> Result<(), String>

Initialises the env_logger for output to a file and to stderr.

This function will create the logfile at file_path if it does not exist, and will truncate it if it does. For further details, see the module docs.

Examples

#[macro_use]
extern crate log;
extern crate maidsafe_utilities;

fn main() {
    assert!(maidsafe_utilities::log::init_to_file(true, "target/test.log").is_ok());
    error!("An error!");
    assert_eq!(maidsafe_utilities::log::init_to_file(true, "target/test.log").unwrap_err(),
        "Logger already initialised.".to_owned());

    // E 22:38:05.499016 <main> [example:main.rs:7] An error!
}