Struct maidsafe_utilities::SeededRng

pub struct SeededRng {
    // some fields omitted
}

A fast pseudorandom number generator for use in tests which allows seeding and prints the seed when the thread in which it is created panics.

Methods

impl SeededRng

fn new() -> Self

Construct a new SeededRng using a seed generated from cryptographically secure random data.

The seed is only set once for the whole process, so every call to this will yield internal RNGs which are all seeded identically.

fn from_seed(seed: [u32; 4]) -> Self

Construct a new SeededRng using seed.

If the underlying static seed has already been initialised to a value different to seed, then this function will panic.

fn new_rng(&mut self) -> XorShiftRng

Construct a new XorShiftRng using a seed generated from random data provided by self.