Struct maidsafe_utilities::SeededRng

pub struct SeededRng {
    // some fields omitted
}

A fast pseudorandom number generator that allows seeding and prints the seed when the thread in which it is created panics.

Methods

impl SeededRng

fn new() -> Self

Construct a new SeededRng using a seed generated from cryptographically secure random data.

fn from_seed(seed: [u32; 4]) -> Self

Construct a new SeededRng using seed.

fn new_rng(&mut self) -> XorShiftRng

Construct a new XorShiftRng using a seed generated from random data provided by self.